requirements.txt
Flask
opencv-python
PyYAML
numpy